# doc/api.yaml
openapi: 3.0.0
paths:
...
  /album:
    get:
      operationId: AlbumController_findAll
      summary: get all albums
      parameters: []
      responses:
        '200':
          description: Albums successfully getted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
      tags:
        - Album
      security:
        - bearer: []
    post:
      operationId: AlbumController_create
      summary: create new album
      parameters: []
      requestBody:
        required: true
        description: Data for new album
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Album successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Request body does not contain required fields
      tags:
        - Album
      security:
        - bearer: []
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      summary: get single album by id
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '200':
          description: Album successfully getted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id not found
      tags:
        - Album
      security:
        - bearer: []
    put:
      operationId: AlbumController_update
      summary: update album info
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      requestBody:
        required: true
        description: Data for update album
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Album successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id not found
      tags:
        - Album
      security:
        - bearer: []
    delete:
      operationId: AlbumController_remove
      summary: delete album
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '204':
          description: Album successfully deleted
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id not found
      tags:
        - Album
      security:
        - bearer: []
...
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbumToFavorites
      summary: add album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '201':
          description: Album successfully added
        '400':
          description: Album id is invalid (not UUID).
        '422':
          description: Album with given id does not exist
      tags:
        - Favorites
      security:
        - bearer: []
    delete:
      operationId: FavoritesController_removeAlbumFromFavorites
      summary: delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Id of album
          schema:
            type: string
      responses:
        '204':
          description: Album successfully deleted
        '400':
          description: Album id is invalid (not UUID)
        '404':
          description: Album with given id does not found
      tags:
        - Favorites
      security:
        - bearer: []
...
  /auth/signup:
    post:
      operationId: AuthController_signup
      summary: send login and password to create a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: >-
            Request body does not contain required fields or the fields are of
            the wrong type
      tags:
        - Auth
  /auth/login:
    post:
      operationId: AuthController_login
      summary: send login and password to get Access token and Refresh token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '200':
          description: Access and refresh tokens successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRefreshTokensPairDto'
        '400':
          description: >-
            Request body does not contain required fields or the fields are of
            the wrong type
        '403':
          description: User with this login does not exist
      tags:
        - Auth
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: >-
        send refresh token in body as { refreshToken } to get new pair of Access
        token and Refresh token
      parameters: []
      requestBody:
        required: true
        description: Refresh token
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Access and refresh tokens successfully refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRefreshTokensPairDto'
        '401':
          description: Refresh token wasn't provided
        '403':
          description: Refresh token is invalid or expired
      tags:
        - Auth
info:
  title: Home Library Service
  description: 'Task #7: Implementation Home Library Service: Part 3'
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        login:
          type: string
          example: login
        version:
          type: number
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: login
        password:
          type: string
          example: password
      required:
        - login
        - password
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - name
        - year
        - artistId
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year
        - artistId
...
    AccessRefreshTokensPairDto:
      type: object
      properties:
        accesshToken:
          type: string
          example: access_token_here
        refreshToken:
          type: string
          example: refresh_token_here
      required:
        - accesshToken
        - refreshToken
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: your_refresh_token_here
      required:
        - refreshToken